pyproject.toml
[tool.poetry]
name = "legisdata"
version = "0.1.0"
description = ""
authors = ["Jeffrey04 <mycyberpet@yahoo.com>"]
readme = "README.md"
packages = [{include = "legisdata", from = "src"}, {include="legisweb", from = "src"}, {include="legisweb_viewer", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
django = "^5.0.7"
django-dotenv = "^1.4.2"
huggingface-hub = {extras = ["cli"], version = "^0.23.4"}
djangorestframework = "^3.15.1"
typedload = "^2.33"
psycopg2-binary = "^2.9.9"
django-structlog = "^8.1.0"
django-typer = {extras = ["rich"], version = "^2.1.2"}
gunicorn = "^22.0.0"
drf-flex-fields = "^1.0.2"
django-opensearch-dsl = "^0.6.2"

[tool.poetry.scripts]
legisdata = 'legisdata.main:app'

[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"

[tool.poetry.group.cli.dependencies]
requests = "^2.32.3"
typer = "^0.12.3"
structlog = "^24.1.0"
unstructured = {extras = ["pdf"], version = "^0.13.7"}
lxml = "^5.2.2"
parsel = "^1.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
